Fix layout for background component/container
