+++ client/components/layout/background/index.js
@@ -1,16 +1,32 @@
-import React from 'react'
+import React, { PropTypes } from 'react'
+import classnames from 'classnames'
 
-// Current module dependencies
 if (process.env.BROWSER) require('./index.scss')
 
-const Background = ({ children }) => (
-  <div className='bg-cover bg-center bg-reboo absolute top-0 right-0 bottom-0 left-0'>
-    <div className='table col-3 mx-auto full-height'>
-      <div className='table-cell align-middle'>
-        {children}
+const Background = ({ children, contentSize, image }) => {
+  const bgStyle = {
+    backgroundImage: image ? `url('${image}')` : undefined,
+    overflow: 'auto'
+  }
+
+  return (
+    <div className='bg-center bg-reboo bg-cover absolute top-0 right-0 bottom-0 left-0' style={bgStyle}>
+      <div className={classnames(`col-${contentSize}`, 'table mx-auto full-height')}>
+        <div className='table-cell align-middle'>
+          {children}
+        </div>
       </div>
     </div>
-  </div>
-)
+  )
+}
+
+Background.propTypes = {
+  contentSize: PropTypes.number,
+  image: PropTypes.string
+}
+
+Background.defaultProps = {
+  contentSize: 3
+}
 
 export default Background
