+++ client/components/layout/background/index.scss
@@ -19,6 +19,10 @@
     color: white;
   }
 
+  h1, h2 {
+    margin: 24px 0;
+  }
+
   a {
     font-weight: 600;
     color: white;
